refactor(slack): extract channel enumeration into channels.ts; nullable enriched fields
Split GET /v1/slack/channels out of the share module into its own route
module (integrations/slack/channels.ts) so non-share behavior stops
piling onto slack/share.ts. Both modules keep their own ROUTES entry in
the shared routes array; share.ts retains only POST /v1/slack/share.
Endpoint URLs, operationIds, and policies are unchanged.

Also switch the enriched channel fields to required-but-nullable:
memberCount, topic, and imageUrl are now `number|string | null` rather
than optional, so every row carries an explicit null when the raw
payload has no value (DMs have null memberCount/topic; channels/groups
have null imageUrl).

Tests split to mirror the module split: channel-listing coverage moves
to slack-channels-routes.test.ts; slack-share-routes.test.ts keeps the
share handler; the token-routing suite still covers both routes since
the read/write auth split is one invariant spanning them.

Part of LUM-2711
